\name{show-methods}
\docType{methods}
\alias{show-methods}
\title{Methods for The Output of \code{whites.htest}}
\description{
This is the \code{show} method of the class \code{whitetest} which is used for \code{whites.htest}.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "whitetest")}}{
It is the method for the printing of the output. 
}
}}

